# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BetPulse API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface AuthCredentials {
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface User {
  id: number;
  email: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface Bet {
  id: number;
  userId: number;
  sport: string;
  betType: string;
  teams: string;
  odds: number;
  stake: number;
  /** @nullable */
  result: string | null;
  /** @nullable */
  profitLoss: number | null;
  /** @nullable */
  notes: string | null;
  createdAt: string;
}

export interface CreateBetBody {
  sport: string;
  betType: string;
  teams: string;
  odds: number;
  stake: number;
  /** @nullable */
  result?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface UpdateBetBody {
  sport?: string;
  betType?: string;
  teams?: string;
  odds?: number;
  stake?: number;
  /** @nullable */
  result?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface Bankroll {
  userId: number;
  startingBankroll: number;
  currentBankroll: number;
  totalProfitLoss: number;
}

export interface UpdateBankrollBody {
  startingBankroll: number;
}

export interface SportBreakdown {
  sport: string;
  totalBets: number;
  wins: number;
  losses: number;
  profitLoss: number;
  winRate: number;
}

export interface Stats {
  totalBets: number;
  wins: number;
  losses: number;
  pushes: number;
  pending: number;
  winRate: number;
  totalStaked: number;
  totalProfitLoss: number;
  roi: number;
  avgOdds: number;
  avgStake: number;
  bySport: SportBreakdown[];
  recentStreak: number;
}

export interface Insights {
  /**
   * @minimum 0
   * @maximum 100
   */
  riskScore: number;
  tiltDetected: boolean;
  tiltExplanation: string;
  chasingLosses: boolean;
  chasingLossesExplanation: string;
  overconfidence: boolean;
  overconfidenceExplanation: string;
  /**
   * @minimum 0
   * @maximum 100
   */
  bankrollDisciplineScore: number;
  improvementRules: string[];
  summary: string;
  generatedAt: string;
}

export type ListBetsParams = {
  sport?: string;
  result?: string;
};