# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BetPulse API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Register new user
 */
export const registerBodyPasswordMin = 6;

export const RegisterBody = zod.object({
  email: zod.string().email(),
  password: zod.string().min(registerBodyPasswordMin),
});

/**
 * @summary Login user
 */
export const loginBodyPasswordMin = 6;

export const LoginBody = zod.object({
  email: zod.string().email(),
  password: zod.string().min(loginBodyPasswordMin),
});

export const LoginResponse = zod.object({
  token: zod.string(),
  user: zod.object({
    id: zod.number(),
    email: zod.string(),
  }),
});

/**
 * @summary Get current user
 */
export const GetMeResponse = zod.object({
  id: zod.number(),
  email: zod.string(),
});

/**
 * @summary List all bets for current user
 */
export const ListBetsQueryParams = zod.object({
  sport: zod.coerce.string().optional(),
  result: zod.coerce.string().optional(),
});

export const ListBetsResponseItem = zod.object({
  id: zod.number(),
  userId: zod.number(),
  sport: zod.string(),
  betType: zod.string(),
  teams: zod.string(),
  odds: zod.number(),
  stake: zod.number(),
  result: zod.string().nullable(),
  profitLoss: zod.number().nullable(),
  notes: zod.string().nullable(),
  createdAt: zod.coerce.date(),
});
export const ListBetsResponse = zod.array(ListBetsResponseItem);

/**
 * @summary Create a new bet
 */
export const CreateBetBody = zod.object({
  sport: zod.string(),
  betType: zod.string(),
  teams: zod.string(),
  odds: zod.number(),
  stake: zod.number(),
  result: zod.string().nullish(),
  notes: zod.string().nullish(),
});

/**
 * @summary Get a single bet
 */
export const GetBetParams = zod.object({
  id: zod.coerce.number(),
});

export const GetBetResponse = zod.object({
  id: zod.number(),
  userId: zod.number(),
  sport: zod.string(),
  betType: zod.string(),
  teams: zod.string(),
  odds: zod.number(),
  stake: zod.number(),
  result: zod.string().nullable(),
  profitLoss: zod.number().nullable(),
  notes: zod.string().nullable(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Update a bet (e.g. set result)
 */
export const UpdateBetParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateBetBody = zod.object({
  sport: zod.string().optional(),
  betType: zod.string().optional(),
  teams: zod.string().optional(),
  odds: zod.number().optional(),
  stake: zod.number().optional(),
  result: zod.string().nullish(),
  notes: zod.string().nullish(),
});

export const UpdateBetResponse = zod.object({
  id: zod.number(),
  userId: zod.number(),
  sport: zod.string(),
  betType: zod.string(),
  teams: zod.string(),
  odds: zod.number(),
  stake: zod.number(),
  result: zod.string().nullable(),
  profitLoss: zod.number().nullable(),
  notes: zod.string().nullable(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Delete a bet
 */
export const DeleteBetParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Get current user's bankroll
 */
export const GetBankrollResponse = zod.object({
  userId: zod.number(),
  startingBankroll: zod.number(),
  currentBankroll: zod.number(),
  totalProfitLoss: zod.number(),
});

/**
 * @summary Set starting bankroll
 */
export const UpdateBankrollBody = zod.object({
  startingBankroll: zod.number(),
});

export const UpdateBankrollResponse = zod.object({
  userId: zod.number(),
  startingBankroll: zod.number(),
  currentBankroll: zod.number(),
  totalProfitLoss: zod.number(),
});

/**
 * @summary Get betting stats summary for current user
 */
export const GetStatsResponse = zod.object({
  totalBets: zod.number(),
  wins: zod.number(),
  losses: zod.number(),
  pushes: zod.number(),
  pending: zod.number(),
  winRate: zod.number(),
  totalStaked: zod.number(),
  totalProfitLoss: zod.number(),
  roi: zod.number(),
  avgOdds: zod.number(),
  avgStake: zod.number(),
  bySport: zod.array(
    zod.object({
      sport: zod.string(),
      totalBets: zod.number(),
      wins: zod.number(),
      losses: zod.number(),
      profitLoss: zod.number(),
      winRate: zod.number(),
    }),
  ),
  recentStreak: zod.number(),
});

/**
 * @summary Get AI behavior insights for current user
 */
export const getInsightsResponseRiskScoreMin = 0;
export const getInsightsResponseRiskScoreMax = 100;

export const getInsightsResponseBankrollDisciplineScoreMin = 0;
export const getInsightsResponseBankrollDisciplineScoreMax = 100;

export const GetInsightsResponse = zod.object({
  riskScore: zod
    .number()
    .min(getInsightsResponseRiskScoreMin)
    .max(getInsightsResponseRiskScoreMax),
  tiltDetected: zod.boolean(),
  tiltExplanation: zod.string(),
  chasingLosses: zod.boolean(),
  chasingLossesExplanation: zod.string(),
  overconfidence: zod.boolean(),
  overconfidenceExplanation: zod.string(),
  bankrollDisciplineScore: zod
    .number()
    .min(getInsightsResponseBankrollDisciplineScoreMin)
    .max(getInsightsResponseBankrollDisciplineScoreMax),
  improvementRules: zod.array(zod.string()),
  summary: zod.string(),
  generatedAt: zod.coerce.date(),
});